Breaking up the fragment writer in hopes it will make this slightly less complicated
