+++ Sources/Buffie/File Writers/HLS/FragmentedMP4InitSegment.swift
@@ -1,0 +1,37 @@
+import Foundation
+import CoreMedia
+
+struct MOOVConfig {
+    
+    var sps: [UInt8]
+    var pps: [UInt8]
+    var width: UInt32
+    var height: UInt32
+    var timescale: UInt32 = 30000
+    
+    init(_ format: CMFormatDescription) {
+        let paramSet = getVideoFormatDescriptionData(format)
+        self.sps = paramSet.first!
+        self.pps = paramSet.last!
+        
+        let dimensions = CMVideoFormatDescriptionGetDimensions(format)
+        self.width     = UInt32(dimensions.width)
+        self.height    = UInt32(dimensions.height)
+    }
+    
+}
+
+class FragementedMP4InitalizationSegment {
+    
+    init(_ file: URL, format: CMFormatDescription, sample: Sample) throws {
+        
+        var config       = MOOVConfig(format)
+        config.timescale = UInt32(sample.duration.timescale)
+        
+        let ftypBytes = try BinaryEncoder.encode(FTYP())
+        let moovBytes = try BinaryEncoder.encode(MOOV(config))
+        
+        let data = Data(bytes: ftypBytes + moovBytes)
+        try data.write(to: file)
+    }
+}
